backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
bcrypt==4.1.3
PyJWT==2.8.0
pydantic[email]==2.7.1
python-multipart==0.0.9
